change to uint64

+++ graph_reader.go
@@ -6,17 +6,17 @@ import (
 )
 
 type JSON_ordering struct {
-	Ordering []uint32 `json:"ordering"`
+	Ordering []uint64 `json:"ordering"`
 }
 
 type JSON_graph struct {
 	Nodes []struct {
 		Color uint16 `json:"color"`
-		ID    uint32 `json:"id"`
+		ID    uint64 `json:"id"`
 	} `json:"nodes"`
 	Links []struct {
-		Source uint32 `json:"source"`
-		Target uint32 `json:"target"`
+		Source uint64 `json:"source"`
+		Target uint64 `json:"target"`
 	} `json:"links"`
 }
 
@@ -33,7 +33,7 @@ func ReadGraph(filename string) graph {
 	dat, _ := LoadJSON[JSON_graph](filename)
 	var output graph = make(graph)
 	for i := 0; i < len(dat.Nodes); i++ {
-		output.AddVertex(uint32(dat.Nodes[i].ID),uint16(dat.Nodes[i].Color))
+		output.AddVertex(uint64(dat.Nodes[i].ID),uint16(dat.Nodes[i].Color))
 	}
 	for i := 0; i < len(dat.Links); i++ {
 		output.AddEdge(dat.Links[i].Source, dat.Links[i].Target)
@@ -41,7 +41,7 @@ func ReadGraph(filename string) graph {
 	return output
 }
 
-func ReadOrdering(filename string) []uint32{
+func ReadOrdering(filename string) []uint64{
 	dat, _ := LoadJSON[JSON_ordering](filename)
 	return dat.Ordering
 }
